FIX: update lessc instructions to fix minifying deprecation warning
lessc gives a warning that the compression option -x is deprecated. This
commit updates the instructions in the readme to install the plugin
less-plugin-clean-css, and use the option --clean-css instead. This
seems to be the recommended solution by upstream lessc. Testing shows
little difference in the resulting file sizes:
lessc -x: 153K
lessc --clean-css: 150K

Fixes: #682